requirements.txt
influxdb-client>=1.38.0
matplotlib>=3.7.0
PyYAML>=6.0
pillow>=10.0.0
pytz>=2023.3
